Fixing mouse selection so that SysMouse isn't selected for all mice
Fixes to SnapCursor() and SetCursorPos() so that both mice can move the cursor. It's hard to calibrate the guncon otherwise.

The offscreen shots function set the shot position to (0,0) when the cursor is at the edges, i.e. when aiming is as far out as possible, since this is unnecessary for games like time crisis that don't utilize offscreen shots an option to disable it is added.